requirements.txt
pandas==2.2.2
python-dateutil==2.9.0.post0
requests==2.32.3
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tqdm==4.66.4
streamlit==1.37.1